requirements.txt
appdirs
asgiref
beautifulsoup4==4.9.3
bs4==0.0.1
certifi
cffi==1.14.5
chardet==4.0.0
coreapi
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
graphene
graphene-django
filelock==3.0.12
idna==2.10
memory-profiler==0.58.0
mysql-connector-python==8.0.24
oauthlib==3.1.1
Pillow==8.2.0
psutil==5.8.0
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.7
